backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pdfplumber
pdf2image
pytesseract
Pillow
python-multipart
